requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
crispy-tailwind==0.5.0
Django==4.2.4
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-js-asset==2.1.0
Pillow==10.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
